working cbm version although slow

+++ share/include/sinlogic.h
@@ -8,7 +8,7 @@ bool testSinIndexRange(uint8_t max);
 void incSinIndex(void);
 int getCurrentSinValue(void);
 void updateSin(speedTableElem_t *slopetable);
-void scaleSin(uint8_t max);
+void scaleSin(int max);
 
 
 #endif // SINLOGIC_H_INCLUDED
